requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7  # Para interagir com o PostgreSQL
python-dotenv==1.0.0